pipeline/requirements.txt
av
numpy
opencv-python
open3d
pycryptodome
python-gnupg
pyyaml
tqdm